Simplify label selection sheet bottom modal interaction
